pyproject.toml
[tool.poetry]
name = "pep8-checker-bot"
version = "0.1.0"
description = "A bot that checks PEP8 compliance on pull requests in GitHub repositories."
authors = ["Mohamed Tahri <simotahri1@gmail.com>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/metahris/pep8-checker-bot"

[tool.poetry.dependencies]
python = ">=3.6"
flake8 = "^5.0.0"
PyGithub = "^1.57.0"

[tool.poetry.dev-dependencies]
pytest = "^7.0"
pytest-mock = "^3.0"

[build-system]
requires = ["poetry>=1.0"]
build-backend = "poetry.masonry.api"
